# docs/swagger/swagger.yaml
basePath: /
definitions:
  gateway.InitiatePaymentRequest:
    properties:
      billing_period:
        enum:
        - monthly
        - annual
        type: string
      cellphone:
        type: string
      email:
        type: string
      methods:
        items:
          type: string
        type: array
      name:
        type: string
      password:
        minLength: 6
        type: string
      tax:
        type: string
    required:
    - billing_period
    - cellphone
    - email
    - methods
    - name
    - password
    - tax
    type: object
  gateway.WebhookBilling:
    properties:
      amount:
        type: integer
      customer:
        $ref: '#/definitions/gateway.WebhookBillingCustomer'
      externalId:
        description: nosso pendingRegistrationID
        type: string
      frequency:
        type: string
      id:
        type: string
      kind:
        items:
          type: string
        type: array
      paidAmount:
        type: integer
      products:
        items:
          $ref: '#/definitions/gateway.WebhookBillingProduct'
        type: array
      status:
        description: '"PAID"'
        type: string
    type: object
  gateway.WebhookBillingCustomer:
    properties:
      id:
        type: string
      metadata:
        additionalProperties:
          type: string
        description: contém email, name, cellphone, taxId
        type: object
    type: object
  gateway.WebhookBillingProduct:
    properties:
      externalId:
        type: string
      id:
        type: string
      quantity:
        type: integer
    type: object
  gateway.WebhookData:
    properties:
      billing:
        $ref: '#/definitions/gateway.WebhookBilling'
      payment:
        $ref: '#/definitions/gateway.WebhookPayment'
    type: object
  gateway.WebhookPayload:
    properties:
      data:
        $ref: '#/definitions/gateway.WebhookData'
      devMode:
        type: boolean
      event:
        description: '"billing.paid", "withdraw.done", etc.'
        type: string
      id:
        type: string
    type: object
  gateway.WebhookPayment:
    properties:
      amount:
        type: integer
      fee:
        type: integer
      method:
        type: string
    type: object
...
  model.CreatePaymentResponse:
    properties:
      url:
        example: https://abacatepay.com/billing/abc123
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      error:
        example: mensagem de erro
        type: string
    type: object
...
  model.Job:
    properties:
      company:
        type: string
      description:
        type: string
      id:
        type: integer
      job_id:
        type: integer
      job_link:
        type: string
      location:
        type: string
      site_id:
        type: integer
      title:
        type: string
    type: object
...
  model.SandboxScrapeErrorResponse:
    properties:
      error:
        example: falha ao executar scraping
        type: string
      message:
        example: erro no scraping
        type: string
      success:
        example: false
        type: boolean
    type: object
  model.SandboxScrapeResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Job'
        type: array
      message:
        example: Scraping concluído com sucesso
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  model.SiteDTO:
    properties:
      base_url:
        example: https://carreiras.ifood.com.br
        type: string
      is_subscribed:
        example: true
        type: boolean
      logo_url:
        example: https://s3.amazonaws.com/logos/ifood.webp
        type: string
      site_id:
        example: 1
        type: integer
      site_name:
        example: iFood
        type: string
    type: object
  model.SiteScrapingConfig:
    properties:
      api_endpoint_template:
        type: string
      api_headers_json:
        type: string
      api_method:
        type: string
      api_payload_template:
        type: string
      base_url:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      job_description_selector:
        type: string
      job_list_item_selector:
        type: string
      job_requisition_id_selector:
        type: string
      json_data_mappings:
        type: string
      link_attribute:
        type: string
      link_selector:
        type: string
      location_selector:
        type: string
      logo_url:
        type: string
      next_page_selector:
        type: string
      scraping_type:
        description: '''HTML'', ''API'', ''HEADLESS'''
        type: string
      site_name:
        type: string
      title_selector:
        type: string
    type: object
  model.StatusResponse:
    properties:
      status:
        example: received
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para plataforma de scraping e matching de vagas de emprego.
  title: ScrapJobs API
  version: "1.0"
paths:
...
  /api/getSites:
    get:
      description: Retorna todos os sites com status de inscricao do usuario
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SiteDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Listar sites de carreiras
      tags:
      - Sites
...
  /api/payments/create/{planId}:
    post:
      consumes:
      - application/json
      description: Inicia processo de pagamento para um plano
      parameters:
      - description: ID do plano
        in: path
        name: planId
        required: true
        type: integer
      - description: Dados de pagamento
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/gateway.InitiatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreatePaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Criar pagamento
      tags:
      - Payment
...
  /api/webhooks/abacatepay:
    post:
      consumes:
      - application/json
      description: Recebe notificacao de pagamento do AbacatePay
      parameters:
      - description: Payload do webhook
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/gateway.WebhookPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Webhook de pagamento
      tags:
      - Payment
...
  /scrape-sandbox:
    post:
      consumes:
      - application/json
      description: Executa scraping de teste com a configuracao fornecida (admin)
      parameters:
      - description: Configuracao de scraping
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.SiteScrapingConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SandboxScrapeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SandboxScrapeErrorResponse'
      security:
      - CookieAuth: []
      summary: Testar scraping (sandbox)
      tags:
      - Sites
  /siteCareer:
    post:
      consumes:
      - multipart/form-data
      description: Cria nova configuracao de scraping para um site (admin)
      parameters:
      - description: JSON da configuracao do site
        in: formData
        name: siteData
        required: true
        type: string
      - description: Logo da empresa
        in: formData
        name: logo
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.SiteScrapingConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Adicionar site de carreiras
      tags:
      - Sites
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: Authorization
    type: apiKey
swagger: "2.0"